refactor: move components to private components folder
